Add files via upload
Updates of the script. Presenter has been modified to present a fixation cross before the main task so that structural and resting state scans can happen while the program is already launched. Eyetracking file has been replaced with the MRC-eyetracking support (requires the MRC eyetracking DLL to be placed in a folder called DLL_1_5_3 in the parentfolder of stim4prf).